requirements.txt
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.25.0
langchain==0.0.267
openai==0.27.8
python-dotenv==1.0.0
joblib==1.3.1
pytest==7.4.0
faiss-cpu==1.7.4
transformers==4.31.0
sentence-transformers==2.2.2
llama-index==0.8.11